+++ group_test.go
@@ -394,5 +394,3 @@ func TestService_DeleteGroup(t *testing.T) {
 		})
 	}
 }
-
-